teamtwo/requirements.txt
gunicorn
django-heroku
requests
python-dotenv
black
flake8
coverage
six
coveralls